+++ pkg/db-locs.ml
@@ -1,5 +1,0 @@
-#!/usr/bin/env ocaml
-#directory "pkg"
-#use "topkg-ext.ml"
-
-let () = Cmd.exec "pkg/db-locs" >>& fun () -> ()