+++ CHANGES.md
@@ -1,5 +1,5 @@
 
-
+- Build depend on topkg.
 - Relicense from BSD3 to ISC.
 
 v0.8.2 2015-08-14 Cambridge (UK)